Resolved conflicts in Makefile.am and examples\Makefile.am as both
branches added items to the EXTRA_DIST list.

+++ src/sysc/communication/sc_prim_channel.h
@@ -238,7 +238,7 @@ public:
 
 
     int size() const
-        { return m_prim_channel_vec.size(); }
+        { return static_cast<int>(m_prim_channel_vec.size()); }
 
     inline void request_update( sc_prim_channel& );
     void async_request_update( sc_prim_channel& );
